comment attachments

+++ src/features/ticket/ticketActions.ts
@@ -2,33 +2,46 @@
 
 import {prisma} from "@/lib/prismaClient";
 import {revalidatePath} from "next/cache";
-import {$Enums} from "@prisma/client";
+import {$Enums, Entity} from "@prisma/client";
 import {cache} from "react";
 import {redirect} from "next/navigation";
 import {z} from "zod";
 import {ActionState, fromErrorToState} from "@/utils/formUtils";
 import {setCookie} from "@/actions/cookies";
 import {fromDollarsToCentsNoMoneyFormat} from "@/utils/currency";
-import {getAuth} from "@/features/auth/authActions";
 import {getAuthOrRedirect, isOwner} from "@/utils/authUtils";
 import {ParsedSearchParams} from "@/features/ticket/search-params";
+import {getActiveMembership, getActiveOrganization} from "@/features/organization/organizationActions";
+import {sizeInMb} from "@/utils/sizeInMb";
+import {ACCEPTED, MAX_SIZE} from "@/features/ticket/constants";
+import {PutObjectCommand} from "@aws-sdk/client-s3";
+import {s3} from "@/lib/aws";
 import TicketStatus = $Enums.TicketStatus;
+import {generateS3Key} from "@/utils/generateS3Key";
 
 
 type getTicketsParams = {
   userId?: string;
   searchParams: Awaited<ParsedSearchParams>;
+  byOrganization?: boolean;
 }
 
-export const getTickets = async ({userId, searchParams}: getTicketsParams) => {
+export const getTickets = async ({userId, searchParams, byOrganization}: getTicketsParams) => {
+
+
+  const activeOrganization = await getActiveOrganization()
+
 
   const where = {
+    organizationId: activeOrganization?.id,
+
+    ...(activeOrganization && byOrganization && {organizationId: activeOrganization.id}),
     userId,
     title: {contains: searchParams.search, mode: 'insensitive' as const}
   }
-  
+
   try {
-   
+
     const [tickets, count] = await prisma.$transaction([
       prisma.ticket.findMany({
           where,
@@ -51,7 +64,7 @@ export const getTickets = async ({userId, searchParams}: getTicketsParams) => {
         where
       })
     ])
-    
+
 
     return {
       list: tickets,
@@ -70,7 +83,7 @@ export const getTickets = async ({userId, searchParams}: getTicketsParams) => {
 
 export const getTicket = cache(async (id: number) => {
   try {
-    return await prisma.ticket.findFirst(
+    const ticket = await prisma.ticket.findFirst(
       {
         where: {
           id: id
@@ -84,6 +97,9 @@ export const getTicket = cache(async (id: number) => {
         }
       }
     )
+
+    return ticket
+
   } catch (err) {
     console.log(err)
     return null
@@ -95,9 +111,17 @@ export const deleteTicket = async (id: number): Promise<ActionState> => {
   const formData = new FormData() // заглушка
 
   const {user} = await getAuthOrRedirect()
+
+  const activeMembership = await getActiveMembership()
+
   if (!isOwner(user, await getTicket(id))) {
     return fromErrorToState(new Error('Not yours!'), formData)
   }
+
+  if (!activeMembership || !activeMembership.canDeleteTicket) {
+    return fromErrorToState(new Error('Не могешь! удалить, нет прав у тебя'), formData)
+  }
+
 
   try {
     await prisma.ticket.delete(
@@ -126,8 +150,8 @@ export const createTicket = async (state: ActionState, formData: FormData): Prom
 
   const {user, activeOrganization} = await getAuthOrRedirect()
   if (!user) return fromErrorToState(new Error('User is not authenticated'), formData)
-  
-  
+
+
   const title = formData.get('title') as string
   const content = formData.get('content') as string
   const deadline = formData.get('deadline') as string
@@ -135,7 +159,12 @@ export const createTicket = async (state: ActionState, formData: FormData): Prom
 
   try {
     const data = updateTicketSchema.parse({title, content, deadline, bounty})
-    const dataForDB = {...data, bounty: fromDollarsToCentsNoMoneyFormat(data.bounty), userId: user.id, organizationId: activeOrganization!.id}
+    const dataForDB = {
+      ...data,
+      bounty: fromDollarsToCentsNoMoneyFormat(data.bounty),
+      userId: user.id,
+      organizationId: activeOrganization!.id
+    }
 
     await prisma.ticket.create(
       {
@@ -204,3 +233,7 @@ export const updateTicketStatus = async (id: number, status: TicketStatus): Prom
     return "ERROR"
   }
 }
+
+
+
+
